+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+### 3.0.3 - 26/08/2021
+
+* Ajout de 2 codes supplémentaires dans le paramètre **format**:
+- **id** : retourne l'identifiant de l’adresse (clef d’interopérabilité)
+- **riv** : retourne le code RIVOLI de l'adresse
+
 ### 3.0.2 - 30/11/2020
 
 * Ajout d'un paramètre 'distance' sur la fonction **get_address**. Ce paramètre permet d'indiquer une distance limite pour retenir ou pas l'adresse (adresse non retenue si le point adresse retrouvé est trop éloigné du point du canevas)